+++ addOns/graphql/CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 - Bug Fixes.
 - Enhanced Support for Script Input Vectors.
 - Options are now exposed through the API.
+- Optional Arguments are enabled by default.
 
 ## [0.1.0] - 2020-08-28
 - First Version